refactor: extract background sdk into plugin
