+++ src/background/store.js
@@ -1,6 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import persistState from '../store/plugins/persistState';
+import sdk from '../store/plugins/sdk';
 import permissions from '../store/modules/permissions';
 import accounts from '../store/modules/accounts';
 import getters from '../store/getters';
@@ -8,7 +9,7 @@ import getters from '../store/getters';
 Vue.use(Vuex);
 
 const store = new Vuex.Store({
-  plugins: [persistState()],
+  plugins: [sdk, persistState()],
   modules: { permissions, accounts },
   getters: {
     'names/getDefault': () => (address) => `placeholder name for ${address}`,
